Stop documenting Python package when compiled w/o Python support
Now that the default is to compile with Python support, there is much
less of a need to build the documentation when M2 is built w/o Python
support.  When this happens, we just use a stub page for the
documentation.

This lets us get rid of the cached examples and the very hacky and
hard-to-maintain file no-python.m2.